# api/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: RESTful API for the Todo application with JWT authentication and multi-user support
  version: 1.0.0
  contact:
    name: Todo App
servers:
  - url: http://localhost:3000/api
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatar_url:
          type: string
          nullable: true
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    UserSettings:
      type: object
      properties:
        mfa_enabled:
          type: boolean
        notifications_enabled:
          type: boolean
        theme:
          type: string
paths:
...
  /user/profile:
    get:
      tags: [User]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

    put:
      tags: [User]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                avatar_url:
                  type: string
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
        '401':
          description: Unauthorized or wrong password
  /user/settings:
    get:
      tags: [User]
      summary: Get user settings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
        '401':
          description: Unauthorized

    put:
      tags: [User]
      summary: Update user settings
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mfa_enabled:
                  type: boolean
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
        '401':
          description: Unauthorized
...